+++ app/routes/components/Button/changelog.codegen.ts
@@ -13,4 +13,5 @@ export const changelog = ''
   + '- 2025-09-29: Make the prop `size` support `xs` and `xl` values & use tokens\n'
   + '  from ThemeProvider.\n'
   + '- 2025-10-04: Add `ghost` variant. Code formatting improvement.\n'
+  + '- 2025-10-08: Add `light` variant, update `ghost` variant styles.\n'
   ;